requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-mercadopago==6.0.1
django-paypal==2.0
idna==3.6
mercadopago==0.3.5
packaging==23.2
pytz==2023.3.post1
requests==2.31.0
setuptools-git==1.2
setuptools-scm==8.0.4
sqlparse==0.4.4
stripe==7.7.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
